+++ src/lib/generator/utils/entity_attribute_types/entity_attribute_from_bigint.dart
@@ -1,13 +1,11 @@
 import 'package:f_orm_m8/f_orm_m8.dart';
 import 'package:f_orm_m8_sqlite/generator/utils/utils.dart';
 
-class EntityAttributeFromBigint extends EntityAttribute {
+class EntityAttributeFromBigint extends EntityAttributeFromNotImplemented {
   EntityAttributeFromBigint(
       String modelTypeName, String modelName, String attributeName,
       {int metadataLevel, List<CompositeConstraint> compositeConstraints})
       : super(modelTypeName, modelName, attributeName,
             metadataLevel: metadataLevel,
-            compositeConstraints: compositeConstraints) {
-    //throw Exception("BigInt is not implemented yet");
-  }
+            compositeConstraints: compositeConstraints) {}
 }