+++ src/lib/generator/utils/entity_attribute_types/entity_attribute_from_duration.dart
@@ -1,7 +1,7 @@
 import 'package:f_orm_m8/f_orm_m8.dart';
 import 'package:f_orm_m8_sqlite/generator/utils/utils.dart';
 
-class EntityAttributeFromDuration extends EntityAttribute {
+class EntityAttributeFromDuration extends EntityAttributeFromNotImplemented {
   EntityAttributeFromDuration(
       String modelTypeName, String modelName, String attributeName,
       {int metadataLevel, List<CompositeConstraint> compositeConstraints})