+++ src/test/validation_test.dart
@@ -8,13 +8,18 @@ import 'package:source_gen_test/source_gen_test.dart';
 
 void main() {
   LibraryReader library_1;
+  LibraryReader library_2;
   final path = testFilePath('test', 'src', 'model');
   final caliber0Path =
       testFilePath('test', 'out', 'multiple_datacolumn_on_field.0.caliber');
+  final caliber1Path =
+      testFilePath('test', 'out', 'type_checker_validation.0.caliber');
 
   setUp(() async {
     library_1 = await initializeLibraryReaderForDirectory(
         path, "bad_multiple_datacolumns_on_field.dart");
+    library_2 = await initializeLibraryReaderForDirectory(
+        path, "type_checkers_probe.dart");
   });
 
   group('Validation tests', () {
@@ -25,6 +30,13 @@ void main() {
       String output = await generator.generate(library_1, null);
 
       final expected = await File(caliber0Path).readAsString();
+      expect(output, expected);
+    });
+
+    test('Test the not Implemented validations field!', () async {
+      String output = await generator.generate(library_2, null);
+
+      final expected = await File(caliber1Path).readAsString();
       expect(output, expected);
     });
   });